CopyParagraphButtonV2 -> CopyParagraphButton
